requirements.txt
aenum==3.1.15
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
deprecated==1.2.17
fastapi==0.115.7
frozenlist==1.5.0
future==1.0.0
h11==0.14.0
idna==3.10
line-bot-sdk==3.14.5
multidict==6.1.0
propcache==0.2.1
pydantic==2.10.6
pydantic-core==2.27.2
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
typing-extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wrapt==1.17.2
yarl==1.18.3
